Moved ToDo integration to its own file. Will need optimization.
