+++ main.py
@@ -1,95 +1,15 @@
 from kivy.app import App
-from kivy.uix.widget import Widget
-from kivy.uix.button import Button
-from kivy.uix.label import Label
-from kivy.uix.boxlayout import BoxLayout
-from kivy.uix.relativelayout import RelativeLayout
-from kivy.uix.checkbox import CheckBox
 from kivy.config import Config
 from kivy.core.window import Window
 from kivy.uix.screenmanager import ScreenManager, Screen
-from kivy.uix.image import AsyncImage
-from integrations.ToDoIntegrations.ToDo import ToDoIntegration
+from integrations.ToDoIntegrations.ToDoWidget import ToDoWidget
 from integrations.WeatherIntegrations.WeatherWidget import WeatherWidget
-from kivy.properties import ObjectProperty
-from kivy.properties import StringProperty
 from kivy.base import runTouchApp
-import threading
-from kivy.clock import Clock
-
-from integrations.ToDoIntegrations.Task import TaskItem
 
 # Set the default size of the window to 480x320, the size of my 3.5" touchscreen module for a Raspberry Pi
 Config.set('graphics', 'width', '480')
 Config.set('graphics', 'height', '320')
 Window.size=(480,320)
-
-# TODO: Move these widget classes to their own files and integrate with the existing classes there
-
-# The widget that handles all transactions for the Microsoft To Do integration.
-class ToDoWidget(BoxLayout):
-    integration = ToDoIntegration()
-    access_code_thread = None
-    sign_in_label_text = "Sign in to Microsoft"
-    tasks = None
-    grid_layout = ObjectProperty()
-
-    def __init__(self, **kwargs):
-        super(ToDoWidget, self).__init__(**kwargs)
-
-        # If an account exists in cache, get it now. If not, don't do anything and let user sign in on settings screen.
-        if(self.integration.app.get_accounts()):
-            self.Get_Access_Code_Threaded()
-
-    #region Authentication
-
-    # Run Get_Access_Code() in a new thread
-    def Get_Access_Code_Threaded(self):
-        access_code_thread = threading.Thread(target=self.Get_Access_Code)
-        access_code_thread.setDaemon(True)
-        access_code_thread.start()
-
-    # Aquire a new access token or pull one from the cache. 
-    # Assuming one was found, pull new task info from the API
-    def Get_Access_Code(self):
-        success = self.integration.Aquire_Access_Token()
-        if success:
-            self.sign_in_label_text = "You are signed in to Microsoft"
-            self.Aquire_Task_Info()
-    
-    #endregion
-
-    # Aquire task information from the To Do integration and render them on screen
-    def Aquire_Task_Info(self):
-        self.tasks = self.integration.Get_Tasks()
-        
-        #This is how it should be able to work. Not sure why this doesn't
-        #grid_layout = self.ids['tasks_list']
-        for task in self.tasks:
-
-            # This is the checkbox item of the new task
-            checkbox = task.children[1]
-            checkbox.bind(active=self.Box_Checked)
-
-            if not task in self.grid_layout.children:
-                print("Adding new task:", task.title)
-                self.grid_layout.add_widget(task)
-
-    def Box_Checked(self, checkbox, value):
-        print(checkbox, "checked with value", value)
-        task = checkbox.parent
-        old_status = task.Get_Status()
-
-        if value:
-            task.Mark_Complete()
-            if old_status != task.Get_Status():
-                self.integration.Update_Task(task)
-                self.Aquire_Task_Info()
-        else:
-            print("Task '", task.Get_Title(), "' is already commplete")
-            task.Mark_Uncomplete()
-            if old_status != task.Get_Status():
-                self.integration.Update_Task(task)
 
 #region Kivy Screens and Manager
